backend/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
sqlmodel>=0.0.8
pydantic>=1.10.7
python-multipart>=0.0.6
pytest>=7.3.1
httpx>=0.24.0
openpyxl>=3.1.2
pandas>=2.0.1
psycopg2-binary>=2.9.6
alembic>=1.10.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4